refactor(views): separate autocomplete views from views.views
